requirements.txt
# Use the PyTorch custom wheel index for CUDA packages
--extra-index-url https://download.pytorch.org/whl/cu124

torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1


--extra-index-url https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax[cuda12_pip]==0.5.3
flax==0.10.4
orbax==0.1.9
optax==0.2.4
optuna==4.2.1



numpy==1.26.4
wandb==0.19.6
matplotlib==3.10.0
hydra-core==1.3.2
pytorch-lightning==2.5.0
holoviews==1.20.0
scipy==1.15.1
eikonalfm==0.9.7
diplib==3.4.1
taichi==1.6.0
scikit-image==0.24.0
openpnm==3.5.0
porespy==2.4.2
pypardiso==0.4.6
plotly==6.0.0
kaleido==0.2.1 
agd==0.2.16